requirements.txt
pytelegrambotapi
python-decouple
requests 
beautifulsoup4 
lxml 
tqdm 